% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buy_units.R
\name{buy_units}
\alias{buy_units}
\title{Buy Units}
\usage{
buy_units(
  infantry = 0,
  artillery = 0,
  tanks = 0,
  fighters = 0,
  bombers = 0,
  aaguns = 0,
  complexes = 0,
  submarines = 0,
  destroyers = 0,
  carriers = 0,
  cruisers = 0,
  battleships = 0
)
}
\arguments{
\item{infantry}{Number of infantry purchased, infantry cost three}

\item{artillery}{Number of artillery purchased, artillery cost four}

\item{tanks}{Number of tanks purchased, tanks cost five}

\item{fighters}{Number of fighters purchased, fighters cost ten}

\item{bombers}{Number of bombers purchased, bombers cost twelve}

\item{aaguns}{Number of anti aircraft guns purchased, anti aircraft guns cost five}

\item{complexes}{Number of industrial complexes purchased, industrial complexes cost fifteen}

\item{submarines}{Number of submarines purchased, submarines cost six}

\item{destroyers}{Number of destroyers purchased, destroyers cost eight}

\item{carriers}{Number of aircraft carriers purchased, aircraft carriers cost fourteen}

\item{cruisers}{Number of cruisers purchased, cruisers cost twelve}

\item{battleships}{Number of battleships purchased, battleships cost twenty}
}
\value{
Numerical cost of purchase
}
\description{
Calculates the cost of a purchase during the purchase units phase
}
\examples{
buy_units(infantry = 3, artillery = 1, tanks = 1, complexes = 1, submarines = 2)
}
