% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_units.R
\name{info_units}
\alias{info_units}
\title{Unit Information}
\usage{
info_units(unit, write_to_console = TRUE)
}
\arguments{
\item{unit}{A unit in axis and allies spring 1942 in all lowercase letters}

\item{write_to_console}{If true, writes the output to the console, if false, returns as a vector}
}
\value{
Describes the unit's attack, defense, movement, and cost, and other details in several lines of text
}
\description{
Gives basic information about the units in axis and allies
}
\examples{
info_units("artillery")
}
