% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raid_battle.R
\name{raid_battle}
\alias{raid_battle}
\title{Strategic Bombing Raid}
\usage{
raid_battle(
  offense_fighters = 0,
  offense_bombers = 1,
  defense_fighters = 0,
  aagun = FALSE,
  write_to_console = TRUE
)
}
\arguments{
\item{offense_fighters}{Number of fighters brought to the strategic bombing raid if using optional rules for strategic bombing raids which include fighters}

\item{offense_bombers}{Number of bombers brought to bombing raid}

\item{defense_fighters}{Number of fighters defending in the strategic bombing raid if using optional rules for strategic bombing raids which include fighters}

\item{aagun}{Is an anti aircraft gun present on the defending side}

\item{write_to_console}{If true, writes the output to the console, if false, returns as a vector}
}
\value{
IPC Damage done by strategic bombing raid to industrial complex, number of offense fighters left if using optional rules, number of bombers left, number of defense fighters left if using optional rules in lines of text
}
\description{
Simulates one strategic bombing raid on an enemy industrial complex
}
\examples{
raid_battle(offense_bombers = 3, aagun = TRUE)
}
