% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea_round.R
\name{sea_round}
\alias{sea_round}
\title{Sea Round}
\usage{
sea_round(
  offense_submarines = 0,
  offense_destroyers = 0,
  offense_carriers = 0,
  offense_cruisers = 0,
  offense_battleships = 0,
  offense_fighters = 0,
  offense_bombers = 0,
  defense_submarines = 0,
  defense_destroyers = 0,
  defense_carriers = 0,
  defense_cruisers = 0,
  defense_battleships = 0,
  defense_fighters = 0,
  write_to_console = TRUE
)
}
\arguments{
\item{offense_submarines}{Number of submarines on the attacking side, which hit when the dice roll is a 2 or less}

\item{offense_destroyers}{Number of destroyers on the attacking side, which hit when the dice roll is a 2 or less}

\item{offense_carriers}{Number of carriers on the attacking side, which hit when the dice roll is a 1}

\item{offense_cruisers}{Number of cruisers on the attacking side, which hit when the dice roll is a 3 or less}

\item{offense_battleships}{Number of battleships on the attacking side, which hit when the dice roll is a 4 or less}

\item{offense_fighters}{Number of fighters on the attacking side, which hit when the dice roll is a 3 or less}

\item{offense_bombers}{Number of bombers on the attacking side, which hit when the dice roll is a 4 or less}

\item{defense_submarines}{Number of submarines on the defending side, which hit when the dice roll is a 1}

\item{defense_destroyers}{Number of destroyers on the defending side, which hit when the dice roll is a 2 or less}

\item{defense_carriers}{Number of carriers on the defending side, which hit when the dice roll is a 2 or less}

\item{defense_cruisers}{Number of cruisers on the defending side, which hit when the dice roll is a 3 or less}

\item{defense_battleships}{Number of battleships on the defending side, which hit when the dice roll is a 4 or less}

\item{defense_fighters}{Number of fighters on the defending side, which hit when the dice roll is a 4 or less}

\item{write_to_console}{If true, writes the output to the console, if false, returns as a vector}
}
\value{
Number of offensive air hits, offensive submarine hits, offensive other hits, defensive air hits, defensive submarine hits, and defensive other hits in several lines of text
}
\description{
Simulates one round of sea combat
}
\examples{
sea_round(offense_submarines = 1, offense_bombers = 1, defense_battleships = 1)
}
